Q&A_knowledge_base/pyproject.toml
[tool.poetry]
name = "local-rag"
version = "0.1.0"
description = "Local-first RAG Q&A with Ollama/OpenRouter + Qdrant"
authors = ["You <you@example.com>"]
packages = [{ include = "lr", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
httpx = "^0.27.0"
pydantic = "^2.8.0"
python-dotenv = "^1.0.1"
qdrant-client = "^1.9.0"
orjson = "^3.10.0"
pypdf = "^4.2.0"
pandas = "^2.2.2"
unstructured = "^0.15.0"
rapidfuzz = "^3.9.0"
# OpenRouter uses the OpenAI-compatible API schema:
openai = "^1.51.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
ruff = "^0.6.0"
ipython = "^8.26.0"

[tool.ruff]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
